Sprint 3 (#591 E5d): split PrivacyHandler into exporters + erasers
Separates the LGPD/GDPR data-handling from the registration controller:
- PrivacyExporters — the six WP personal-data exporters (profile, certificates,
  appointments, audience groups, audience bookings, usermeta), moved verbatim
  with the ITEMS_PER_PAGE const + DatabaseHelperTrait.
- PrivacyErasers — erase_personal_data.
- PrivacyHandler (878 → 156) keeps init, register_exporters/register_erasers,
  and the privacy-policy content; its registration callbacks now point at the
  new classes (WP invokes by callable — behavior identical).

Exporter/eraser return shapes, pagination, queries and decryption are unchanged.
Tests repointed to the new classes + @covers (verified exporters 339/359,
erasers 100/102). Behavior-preserving — Privacy suite green (58 tests);
PHPStan L8 + WPCS clean.
